+++ kleros-app/package.json
@@ -1,0 +1,63 @@
+{
+  "name": "@kleros/kleros-app",
+  "version": "1.0.0",
+  "source": "src/lib/index.ts",
+  "main": "dist/main.js",
+  "module": "dist/module.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "clear": "rm -r ../.parcel-cache",
+    "clean": "rm -rf dist",
+    "start": "parcel src/index.html",
+    "build": "yarn clear & yarn clean & yarn parcel build",
+    "check-style": "eslint 'src/**/*.{ts,tsx}' -fix",
+    "check-types": "tsc --noEmit"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/kleros/kleros-v2.git"
+  },
+  "keywords": [
+    "kleros",
+    "dapp",
+    "atlas"
+  ],
+  "author": "Kleros",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/kleros/kleros-v2/issues"
+  },
+  "homepage": "https://github.com/kleros/kleros-v2#readme",
+  "description": "",
+  "files": [
+    "dist"
+  ],
+  "prettier": "@kleros/kleros-v2-prettier-config",
+  "devDependencies": {
+    "@eslint/compat": "^1.2.2",
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.14.0",
+    "@kleros/kleros-v2-eslint-config": "workspace:^",
+    "@kleros/kleros-v2-prettier-config": "workspace:^",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "eslint": "^9.14.0",
+    "eslint-config-prettier": "^9.1.0",
+    "globals": "^15.12.0",
+    "parcel": "^2.12.0",
+    "typescript": "^5.6.3"
+  },
+  "dependencies": {
+    "@kleros/ui-components-library": "^2.15.0",
+    "jose": "^5.9.6",
+    "wagmi": "^2.13.0"
+  },
+  "peerDependencies": {
+    "@tanstack/react-query": "^5.59.20",
+    "graphql": "^16.9.0",
+    "graphql-request": "^7.1.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "viem": "^2.21.42"
+  }
+}
