+++ kleros-app/src/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Kleros App</title>
+  </head>
+
+  <body>
+    <div id="app"></div>
+    <script type="module" src="App.tsx"></script>
+  </body>
+</html>
