+++ kleros-app/src/lib/atlas/providers/index.ts
@@ -1,0 +1,1 @@
+export * from "./AtlasProvider";
