+++ kleros-app/src/lib/atlas/utils/confirmEmail.ts
