+++ package.json
@@ -25,7 +25,8 @@
     "web-devtools",
     "eslint-config",
     "prettier-config",
-    "tsconfig"
+    "tsconfig",
+    "kleros-app"
   ],
   "packageManager": "yarn@4.5.1",
   "volta": {
@@ -35,6 +36,8 @@
   "devDependencies": {
     "@commitlint/cli": "^17.8.1",
     "@commitlint/config-conventional": "^17.8.1",
+    "@parcel/packager-ts": "2.13.0",
+    "@parcel/transformer-typescript-types": "2.13.0",
     "assert": "^2.0.0",
     "buffer": "^5.7.1",
     "conventional-changelog-cli": "^2.2.2",
@@ -44,7 +47,8 @@
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.0",
     "process": "^0.11.10",
-    "string_decoder": "^1.3.0"
+    "string_decoder": "^1.3.0",
+    "typescript": ">=3.0.0"
   },
   "resolutions": {
     "async@npm^2.4.0": "^2.6.4",
