+++ web/src/components/EnsureAuth.tsx
@@ -1,10 +1,9 @@
-import React from "react";
+import React, { useCallback } from "react";
 
 import { useAccount } from "wagmi";
-
+import { useAtlasProvider } from "@kleros/kleros-app";
 import { Button } from "@kleros/ui-components-library";
-
-import { useAtlasProvider } from "context/AtlasProvider";
+import { errorToast, infoToast, successToast } from "utils/wrapWithToast";
 
 interface IEnsureAuth {
   children: React.ReactElement;
@@ -14,12 +13,22 @@ interface IEnsureAuth {
 const EnsureAuth: React.FC<IEnsureAuth> = ({ children, className }) => {
   const { address } = useAccount();
   const { isVerified, isSigningIn, authoriseUser } = useAtlasProvider();
+  const handleClick = useCallback(() => {
+    infoToast(`Signing in User...`);
+
+    authoriseUser()
+      .then(() => successToast("Signed In successfully!"))
+      .catch((err) => {
+        console.log(err);
+        errorToast(`Sign-In failed: ${err?.message}`);
+      });
+  }, [authoriseUser]);
   return isVerified ? (
     children
   ) : (
     <Button
       text="Sign In"
-      onClick={authoriseUser}
+      onClick={handleClick}
       disabled={isSigningIn || !address}
       isLoading={isSigningIn}
       {...{ className }}
