+++ web/src/layout/Header/navbar/Menu/Settings/Notifications/FormContactDetails/index.tsx
@@ -6,7 +6,7 @@ import { useAccount } from "wagmi";
 import { Button } from "@kleros/ui-components-library";
 
 import { EMAIL_REGEX } from "consts/index";
-import { useAtlasProvider } from "context/AtlasProvider";
+import { useAtlasProvider } from "@kleros/kleros-app";
 
 import { responsiveSize } from "styles/responsiveSize";
 
@@ -17,6 +17,7 @@ import FormContact from "./FormContact";
 import { isUndefined } from "src/utils";
 import InfoCard from "components/InfoCard";
 import { timeLeftUntil } from "utils/date";
+import { errorToast, infoToast, successToast } from "utils/wrapWithToast";
 
 const FormContainer = styled.form`
   width: 100%;
@@ -75,24 +76,34 @@ const FormContactDetails: React.FC<ISettings> = ({ toggleIsSettingsOpen }) => {
       const data = {
         newEmail: emailInput,
       };
+      infoToast("Updating Email ...");
       updateEmail(data)
         .then(async (res) => {
           if (res) {
+            successToast("Email Updated successfully!");
             toggleIsSettingsOpen();
           }
         })
-        .catch((err) => console.log(err));
+        .catch((err) => {
+          console.log(err);
+          errorToast(`Updating Email failed: ${err?.message}`);
+        });
     } else {
       const data = {
         email: emailInput,
       };
+      infoToast("Adding User ...");
       addUser(data)
         .then(async (res) => {
           if (res) {
+            successToast("User added successfully!");
             toggleIsSettingsOpen();
           }
         })
-        .catch((err) => console.log(err));
+        .catch((err) => {
+          console.log(err);
+          errorToast(`Adding User failed: ${err?.message}`);
+        });
     }
   };
