+++ web/src/pages/Settings/EmailConfirmation/index.tsx
@@ -10,7 +10,7 @@ import CheckIcon from "svgs/icons/check-circle-outline.svg";
 import WarningIcon from "svgs/icons/warning-outline.svg";
 import InvalidIcon from "svgs/label-icons/minus-circle.svg";
 
-import { useAtlasProvider } from "context/AtlasProvider";
+import { useAtlasProvider } from "@kleros/kleros-app";
 
 import { landscapeStyle } from "styles/landscapeStyle";
 
@@ -59,10 +59,9 @@ const textCss = css`
 
 const Header = styled.h1<{ fontColor: string }>`
   ${textCss}
-  ${({ fontColor }) =>
-    css`
-      color: ${({ theme }) => theme[fontColor]};
-    `};
+  ${({ fontColor }) => css`
+    color: ${({ theme }) => theme[fontColor]};
+  `};
 `;
 
 const Subtitle = styled.h3`
@@ -74,12 +73,11 @@ const HeaderIconContainer = styled.div<{ iconColor: string }>`
   svg {
     width: 64px;
     height: 64px;
-    ${({ iconColor }) =>
-      css`
-        path {
-          fill: ${({ theme }) => theme[iconColor]};
-        }
-      `}
+    ${({ iconColor }) => css`
+      path {
+        fill: ${({ theme }) => theme[iconColor]};
+      }
+    `}
   }
 `;
