--- web/src/utils/atlas/index.ts
@@ -1,8 +1,0 @@
-export * from "./loginUser";
-export * from "./getNonce";
-export * from "./createMessage";
-export * from "./addUser";
-export * from "./fetchUser";
-export * from "./updateEmail";
-export * from "./confirmEmail";
-export * from "./uploadToIpfs";
